move remaining logi to plotting._matplotlib
