+++ src/pymovements/plotting/scanpathplot.py
@@ -31,10 +31,8 @@ from matplotlib.patches import Circle
 
 from pymovements.events import EventDataFrame
 from pymovements.gaze import GazeDataFrame
-from pymovements.utils.plotting import draw_line_data
-from pymovements.utils.plotting import LinearSegmentedColormapType
-from pymovements.utils.plotting import setup_matplotlib
-
+from pymovements.plotting._matplotlib import LinearSegmentedColormapType, setup_matplotlib, \
+    draw_line_data
 
 # This is really a dirty workaround to use the Agg backend if runnning pytest.
 # This is needed as Windows workers on GitHub fail randomly with other backends.
