+++ src/pymovements/utils/plotting.py
@@ -17,85 +17,24 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-"""Functions for plotting."""
+"""Functions for plotting.
+
+.. deprecated:: v0.21.1
+   This module will be removed in v0.26.0.
+"""
 from __future__ import annotations
 
-from collections.abc import Sequence
-from typing import Literal
-from typing import Union
-
+import matplotlib.colors
 import matplotlib.pyplot
-import numpy as np
-import PIL.Image
 from deprecated.sphinx import deprecated
-from matplotlib.collections import LineCollection
-from typing_extensions import TypeAlias
 
 from pymovements.stimulus.image import _draw_image_stimulus
 
 
-LinearSegmentedColormapType: TypeAlias = dict[
-    Literal['red', 'green', 'blue', 'alpha'],
-    Sequence[tuple[float, ...]],
-]
-
-
-DEFAULT_SEGMENTDATA: LinearSegmentedColormapType = {
-    'red': [
-        (0.0, 0.0, 0.0),
-        (0.5, 1.0, 1.0),
-        (1.0, 1.0, 1.0),
-    ],
-    'green': [
-        (0.0, 0.0, 0.0),
-        (0.5, 1.0, 1.0),
-        (1.0, 0.0, 0.0),
-    ],
-    'blue': [
-        (0.0, 0.0, 0.0),
-        (0.5, 0.0, 0.0),
-        (1.0, 0.0, 0.0),
-    ],
-}
-
-
-DEFAULT_SEGMENTDATA_TWOSLOPE: LinearSegmentedColormapType = {
-    'red': [
-        (0.0, 0.0, 0.0),
-        (0.5, 0.0, 0.0),
-        (0.75, 1.0, 1.0),
-        (1.0, 1.0, 1.0),
-    ],
-    'green': [
-        (0.0, 0.0, 0.0),
-        (0.25, 1.0, 1.0),
-        (0.5, 0.0, 0.0),
-        (0.75, 1.0, 1.0),
-        (1.0, 0.0, 0.0),
-    ],
-    'blue': [
-        (0.0, 1.0, 1.0),
-        (0.25, 1.0, 1.0),
-        (0.5, 0.0, 0.0),
-        (1.0, 0.0, 0.0),
-    ],
-}
-
-CmapNormType: TypeAlias = Union[
-    matplotlib.colors.TwoSlopeNorm,
-    matplotlib.colors.Normalize,
-    matplotlib.colors.NoNorm,
-]
-MatplotlibSetupType: TypeAlias = tuple[
-    matplotlib.pyplot.figure,
-    matplotlib.pyplot.Axes,
-    matplotlib.colors.Colormap,
-    CmapNormType,
-    np.ndarray,
-    bool,
-]
-
-
+@deprecated(
+    reason='This function will be removed in v0.26.0.',
+    version='v0.21.1',
+)
 def setup_matplotlib(
         x_signal: np.ndarray,
         y_signal: np.ndarray,
@@ -112,6 +51,9 @@ def setup_matplotlib(
         pad_factor: float | None = 0.05,
 ) -> MatplotlibSetupType:
     """Configure cmap.
+
+    .. deprecated:: v0.21.1
+       This function will be removed in v0.26.0.
 
     Parameters
     ----------
@@ -147,77 +89,21 @@ def setup_matplotlib(
     MatplotlibSetupType
         Configures fig, ax, cmap, cmap_norm, cmap_segmentdata, cval, and show_cbar.
     """
-    n = len(x_signal)
-
-    fig = matplotlib.pyplot.figure(figsize=figsize)
-    ax = fig.gca()
-
-    if add_stimulus:
-        img = PIL.Image.open(path_to_image_stimulus)
-        ax.imshow(img, origin=stimulus_origin, extent=None)
-    else:
-        if padding is None:
-            x_pad = (np.nanmax(x_signal) - np.nanmin(x_signal)) * pad_factor
-            y_pad = (np.nanmax(y_signal) - np.nanmin(y_signal)) * pad_factor
-        else:
-            x_pad = padding
-            y_pad = padding
-
-        ax.set_xlim(np.nanmin(x_signal) - x_pad, np.nanmax(x_signal) + x_pad)
-        ax.set_ylim(np.nanmin(y_signal) - y_pad, np.nanmax(y_signal) + y_pad)
-        ax.invert_yaxis()
-
-    if cval is None:
-        cval = np.zeros(n)
-        show_cbar = False
-
-    cval_max = np.nanmax(np.abs(cval))
-    cval_min = np.nanmin(cval).astype(float)
-
-    if cmap_norm is None:
-        if cval_max and cval_min < 0:
-            cmap_norm = 'twoslope'
-        elif cval_max:
-            cmap_norm = 'normalize'
-        else:
-            cmap_norm = 'nonorm'
-
-    if cmap is None:
-        if cmap_segmentdata is None:
-            if cmap_norm == 'twoslope':
-                cmap_segmentdata = DEFAULT_SEGMENTDATA_TWOSLOPE
-            else:
-                cmap_segmentdata = DEFAULT_SEGMENTDATA
-
-        cmap = matplotlib.colors.LinearSegmentedColormap(
-            'line_cmap', segmentdata=cmap_segmentdata, N=512,
-        )
-
-    if cmap_norm == 'twoslope':
-        cmap_norm = matplotlib.colors.TwoSlopeNorm(
-            vcenter=0, vmin=-cval_max, vmax=cval_max,
-        )
-    elif cmap_norm == 'normalize':
-        cmap_norm = matplotlib.colors.Normalize(
-            vmin=cval_min, vmax=cval_max,
-        )
-    elif cmap_norm == 'nonorm':
-        cmap_norm = matplotlib.colors.NoNorm()
-
-    elif isinstance(cmap_norm, str):
-        # pylint: disable=protected-access
-
-        # to handle after https://github.com/pydata/xarray/pull/8030 is merged
-        if (
-            scale_class := matplotlib.scale._scale_mapping.get(cmap_norm, None)  # type: ignore
-        ) is None:
-            raise ValueError(f'cmap_norm string {cmap_norm} is not supported')
-
-        norm_class = matplotlib.colors.make_norm_from_scale(scale_class)
-        cmap_norm = norm_class(matplotlib.colors.Normalize)()
-
-    return fig, ax, cmap, cmap_norm, cval, show_cbar
-
+    return _setup_matplotlib(
+        x_signal=x_signal,
+        y_signal=y_signal,
+        figsize=figsize,
+        cmap=cmap,
+        cmap_norm=cmap_norm,
+        cmap_segmentdata=cmap_segmentdata,
+        cval=cval,
+        show_cbar=show_cbar,
+        add_stimulus=add_stimulus,
+        path_to_image_stimulus=path_to_image_stimulus,
+        stimulus_origin=stimulus_origin,
+        padding=padding,
+        pad_factor=pad_factor,
+    )
 
 @deprecated(
     reason='Please use ImageStimulus.show() instead. '
@@ -272,6 +158,10 @@ def draw_image_stimulus(
     )
 
 
+@deprecated(
+    reason='This function will be removed in v0.26.0.',
+    version='v0.21.1',
+)
 def draw_line_data(
         x_signal: np.ndarray,
         y_signal: np.ndarray,
@@ -281,6 +171,9 @@ def draw_line_data(
         cval: np.ndarray | None = None,
 ) -> matplotlib.pyplot.Axes:
     """Draw line data.
+
+    .. deprecated:: v0.21.1
+       This function will be removed in v0.26.0.
 
     Parameters
     ----------
@@ -301,18 +194,12 @@ def draw_line_data(
     -------
     matplotlib.pyplot.Axes
         Axes with added line data.
-
     """
-    # Create a set of line segments so that we can color them individually
-    # This creates the points as a N x 1 x 2 array so that we can stack points
-    # together easily to get the segments. The segments array for line collection
-    # needs to be (numlines) x (points per line) x 2 (for x and y)
-    points = np.array([x_signal, y_signal]).T.reshape((-1, 1, 2))
-    segments = np.concatenate([points[:-1], points[1:]], axis=1)
-    # Create a continuous norm to map from data points to colors
-    line_collection = LineCollection(segments, cmap=cmap, norm=cmap_norm)
-    # Set the values used for colormapping
-    line_collection.set_array(cval)
-    line_collection.set_linewidth(2)
-    line = ax.add_collection(line_collection)
-    return line
+    return _draw_line_data(
+        x_signal=x_signal,
+        y_signal=y_signal,
+        ax=ax,
+        cmap=cmap,
+        cmap_norm=cmap_norm,
+        cval=cval,
+    )
